refactor: 그리디 실전문제 폴더 이동
